Changed URL parameters for retrieving one champion
Instead of passing in the champion ID in the URL, I decided to pass the champion name instead, which is filtered to only have letters from the alphabet. This is also the reason for utils/championNameCounterParts.js where each passed filtered champion name has a corresponding value which is the REAL champion name with spaces to be used in retrieving data from the database.

REFACTORED UI

Changed Color Scheme of website

Refactored design for Player Build page. Currently editing Champion Builds page.

Added react-tabs package for displaying champion counters in Champion Builds Page